Fix inconsistency in websocket example
Previously, when sending text, the input box was not cleaned up.
When clicking Send button again, no text is sent to echo server and
nothing added to the list. This was caused by the model being updated,
but not the view, since data is taken directly from input (newInput),
while not rendered back in the view, thus creating an inconsitency
between view and model